+++ docs/whats_new.md
@@ -5,6 +5,7 @@
 * Now initialize all class attributes with None and removed usage of `hasattr` which simplifies and clarifies some code.
 * Improved handling of `start_time`, `end_time`, `duration`, and `steps` in `manager.py` which fixed a bug in which users couldn't input `start_time` and have the simulation run successfully.
 * simplified handling of `horizontal_diffusivity` in `opendrift` model.
+* user can change `end_time`, `duration`, and `steps` and have the others update accordingly. Tests added to check this.
 
 
 ## v0.6.0 (February 15, 2024)